> Hercules merges:
* e9d07bd:
Fixed bug with garments not displaying properly. (bugreport:7018)

* 9f75e12:
Fixed Ankle Snare to behave as in official servers, you only stop walking when you reach your destination and activates all other traps in the path. (bugreport:4832)

* cba3bef:
New skill config - After due consideration we thought it'd be in the best interest of the community to provide a switch to this official feature since it drastically change the way traps can be used.